Redirect existing DNS servers instead of adding one
This prevents Android from falling back on one of those and
thus circumvent our ad protection. As a bonus, it also removes
the need for us to implement fallback: We just forward the OK
packets to the original server now, and Android can take care
of fallback as usual.

We need to add the server addresses to the builder too, otherwise
the route seems to be ignored.