Do not ignore InterruptedException during sleep()
When we are sleeping while trying to reconnect, and want to stop
the thread, we would end up interrupting the sleep, but that would
be ignored and the thread would never exit.

I'm not sure how this got in there, it does not seem to exist in
the original AdBuster code. This probably helps with issue #9.