Read from DNS sockets first before interacting with device
The DNS socket queue is bounded by both size (1000 items) and time
(10 seconds). If we read from the device first, and insert a new
item into that queue, it might evict the oldest item - even if that
is an item we still want to read from.

This might be slightly sub-optimal: We now can't use the time we
spend on reading DNS responses for transmitting new DNS requests
as before.