Fix unclosed sockets for blacklisted hosts
We called handleDnsRequest with a socket, regardless of whether
we actually need a socket or not. In case we did not, we never
closed it.

Move the creation of the socket to the point where we need it
and also take care of properly cleaning it up in case an error
occurs.