Do not crash on UDP packets without a payload
Firefox sends empty UDP packets to the default gateway to work
around WiFi powersaving, as discussed in:

 https://bugzilla.mozilla.org/show_bug.cgi?id=888268

Our code expects that a UDP packet has a payload, and thus crashes
with NullPointerException when calling getRawData() on the payload.

There are two solutions to this problem: Ignore those packets, as
we do with others, or just decide to send an empty UDP packet to
the same destination. This commit does the latter, as we don't
want to prevent Firefox from its workaround. Note that the source
port will be different, so a server won't be able to respond to
a packet, as we destroy the socket after sending.

Fixes #48