AdVpnService: Do not crash if activity did not run earlier
The code here was reading MainAcitivity.config, which was set
by MainActivity. Just load it as we do in the thread code for
now.

This fixes #51, but is slightly suboptimal: It can still crash
in case the config file is incompatible (because it tries to make
a Toast, but can't since we're not in a running Looper thread thingy).
But we have the same risk in the other code - fixing it will require
a larger reworking.