Do not break other connections to DNS servers, map instead
Instead of directly adding the upstream DNS servers to our VPN,
choose a subnet that we can control and map hosts in that subnet
to the upstream DNS servers.

This allows us to implement subnet.2 - subnet.254 as DNS servers,
allowing us to use 253 DNS servers in total. More DNS servers
will be ignored or cause an error.

As the subnet, we need to make sure not to overlap with used
subnets. The IETF has reserved several ranges for documentation
purposes, so we assume they are free and use them as a hack. If
Android forbids that, we fall back to the previous behavior.

Using the reserved IP addresses here is really evil, and I feel
very bad about it, but it seems more likely to not cause issues
than what we currently use...

Fixes #49