Whitelist: Load app icons asynchronously on-demand
Loading all the app icons upfront causes a huge memory
use (40MB or so) and means that the list takes longer
to load. By loading the images on demand, we only need
drawables for (roughly) the currently displayed items,
saving a ton of space.

We load the app icons asynchronously to not block the
UI thread - the easy approach of loading the icons in
the getView() caused some minor stuttering.

Finally, we keep a weak reference on the drawables inside
the entries. This means that if scrolling down and up again,
the system can directly re-use drawables that have not been
collected yet and does not even to run an async task at all,
making performance even smoother, and improving memory use.