Introduce a new, more clever, host database
In the previous solution, all hosts lists are re-read whenever
the service is restarted, or when the connection is changed. This
is wasteful: Most of the time it does not have to change.

The new database caches read files and only reloads those files
that need reloading. The database is fully atomic, it presents
either the old or the new state.