Use a single VPN thread and interact with commands
Previously, we would shutdown the thread and start a new one,
leading to somewhat of a mess. Let's bind threads to service
in a one-to-one relationship instead, that is, there is exactly
one thread for the lifetime of the service.

Instead of shutting down the thread, we repurpose the block
and interrupt pipe into a command pipe: Single-byte commands
can be written to it, and they are processed in an iteration
of the mainloop.

This also gets rid of the retries on connect failure and simplifies
the code a great deal.