AdVpnService: Unset vpnThread in onDestroy
In theory, the vpn service should be garbage collected, and thus
the thread with it. In practice it turns out not to happen, because
Android's VpnService implementation returns an object of an inner
class in it's onBind() method:

https://android.googlesource.com/platform/frameworks/base/+/master/core/java/android/net/VpnService.java

So, in order to clean up as far as possible, unset the thread when
we are not using it anymore.