Whitelist: Only call setChecked after setting checked listener
If we call setChecked() before setting the listener, we could
be setting it on a re-used view that still has the old listener
set. This would cause an item to get the state from another item.

Also change the listener to do nothing if the state is the same,
this avoids unneccessary config rewrites.

This fixes #92, a regression introduced in commit
030545b188a13263da770fed654a4e92d0cb1bab.