Update the database in refresh instead of files
Update the database in the background, showing a progress
dialog in the foreground; and update the database directly,
rather than using the download manager to download host files
to the system.

This means we do not need WRITE_EXTERNAL_STORAGE permission
anymore to hack around the DownloadManager needing that in
older Android versions, so remove that permission from the
set.

Fixes #43