DnsPacketProxy: Allow clients to cache blocks for 5 seconds
This eliminates duplicates queries when loading a webpage, which
might have a slight improvement in performance. On one of my test
pages, the number of queries reduced from 46 to 19.

The time out for the record needs to be chosen so that unblocking
a host still works "immediately", that is, the time out needs to
be lower than the time needed to unblock a host. It also needs to
be large enough to be effective over a longer page load. 5 seconds
seems like a good compromise: While it currently takes longer to
enter a host into the list, I imagine we will soon-ish have the
ability to unblock hosts with a single click in some "recent
requests" view - it seems like 5 seconds would still work for
that.

The other values are extremely fake. The host names and email
addresses are all set to an invalid dns66.dns66.invalid., and
the serial number and other integer fields are set to 0. This
seems to work - I have not seen any negative effect yet.