New color scheme and round icon
Switch from Indigo and Pink to Teal and Light Blue. Create a new
icon from Android Studio that is round and uses the teal background,
with the DNS66 word in the center set in Cantarell.

This fixes #10, in the sense that we now have round icons. No icon
for quick settings toggles yet, though.