Adjust paddings in extra bar expand button and switches
Make the expand icon a bit larger, by reducing the padding, and
on the left side filling it up with a margin to keep it aligned
with the rest of the activity.

Also adjust the switches to make them easier to hit: A switch
inside the extra bar has 8dp padding on top and bottom, a switch
at the end has 16dp at the bottom, and the global switch has 16dp
on both sides.