ItemActivity: Add attach_file drawable to location textview
Allow users to select a file by clicking the drawable, rather
than having to use a hideous overflow menu. Thanks to the
drawable support of TextView, this actually looks quite good.

Tinting is a bit annoying. API level 23 has an attribute for
that, but our minimum is 21, so we have to do it the hard
way.