Change text appearance and icon alpha in item rows
Previously, the title of an item just used bold text at a larger
size, and the location was in normal design. Use the appropriate
text appearance attributes instead.

Also set the alpha of the image, instead of the view, this should
be more efficient - and while at it, change the alpha to be larger
so it matches the new title apperance, and choose a low alpha for
ignored items.