Explicitly allow traffic for IPv4 and IPv6
On networks where DNS servers are only configured for one
version of the internet protocol, we only add DNS servers
to the VPN corresponding to these. Android then blocks all
traffic on the other protocol version - which makes sense
for a VPN (because you don't want leaks), but does not make
much sense for a host blocker.

Fixes #129