Re-add external storage permissions
Some vendors like Xiaomi seem to produce horribly broken
systems that require external storage permission for
ACTION_CREATE_DOCUMENT with CATEGORY_OPENABLE. They
just pop up a toast saying to toggle on file permissions
in settings, but if we do not list file permissions in
the manifest, there are none to be toggled on in settings.

This does not really have any effect on other Android >= 6.0
phone, as permissions have to be requested at run-time and the
manifest just specifies an upper set, but it should allow Mi
users to work around their horribly broken systems.